Cleanup logic in class_diff::ensure_lookup_tables_populated()
At some point, it appeared that some bogus DWARF wouldn't tie a
function decl with its underlying symbol, but subsequent version of
the DWARF emitter (in the second subject of the diff) would correctly
link the function decl with its underlying symbol.

Comparing the two versions of function decl could then wrongly make
the class_diff code think that the function decl was added to the
binary.  I later added code that checks that for every supposedly
added function, its symbol must *NOT* have been present in the first
version of the binary.  I added some similar code for the removed
symbols case.  And that added code seems to work OK.

But then there is an even more ancient hacky attempt at handling the
same case that is no more warranted.  This patch removes it.

	* src/abg-comparison.cc
	(class_diff::ensure_lookup_tables_populated): Remove the code that
	tries to lookup allegedly added functions from the set of deleted
	ones, by using the pretty printed name of the function.  Handling
	the case of a function decl not correctly tied to it symbol is
	handled my generically a bit later in this function.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>